+++ Source/MouseManager.cpp
@@ -48,7 +48,7 @@ bool MouseManager::ButtonReleased(const Buttons& button) {
 }
 
 MouseManager::Buttons MouseManager::GetButton() {
-	for (int i = 0; i < 5; i++) {
+	for (size_t i = 0; i < 5; i++) {
 		Buttons currentButton = static_cast<Buttons>(i);
 		if(ButtonDown(currentButton)) return currentButton;
 	}
